Latest/requirements.txt
flask==3.0.0
gunicorn==21.2.0
twilio==9.0.0
google-genai==0.2.2
python-dotenv==1.0.0
